+++ packages/node/src/integrations/linkederrors.ts
@@ -1,6 +1,5 @@
-import { addGlobalEventProcessor, getCurrentHub } from '@sentry/core';
-import type { Event, EventHint, Exception, ExtendedError, Integration, StackParser } from '@sentry/types';
-import { isInstanceOf } from '@sentry/utils';
+import type { Event, EventHint, EventProcessor, Hub, Integration } from '@sentry/types';
+import { applyAggregateErrorsToEvent } from '@sentry/utils';
 
 import { exceptionFromError } from '../eventbuilder';
 import { ContextLines } from './contextlines';
@@ -41,14 +40,24 @@ export class LinkedErrors implements Integration {
   /**
    * @inheritDoc
    */
-  public setupOnce(): void {
-    addGlobalEventProcessor(async (event: Event, hint: EventHint) => {
+  public setupOnce(addGlobalEventProcessor: (callback: EventProcessor) => void, getCurrentHub: () => Hub): void {
+    addGlobalEventProcessor(async (event: Event, hint?: EventHint) => {
       const hub = getCurrentHub();
-      const self = hub.getIntegration(LinkedErrors);
       const client = hub.getClient();
-      if (client && self && self._handler && typeof self._handler === 'function') {
-        self._handler(client.getOptions().stackParser, event, hint);
+      const self = hub.getIntegration(LinkedErrors);
+
+      if (!client || !self) {
+        return event;
       }
+
+      applyAggregateErrorsToEvent(
+        exceptionFromError,
+        client.getOptions().stackParser,
+        self._key,
+        self._limit,
+        event,
+        hint,
+      );
 
       // If the ContextLines integration is enabled, we add source code context to linked errors
       // because we can't guarantee the order that integrations are run.
@@ -59,36 +68,5 @@ export class LinkedErrors implements Integration {
 
       return event;
     });
-  }
-
-  /**
-   * @inheritDoc
-   */
-  private _handler(stackParser: StackParser, event: Event, hint: EventHint): Event {
-    if (!event.exception || !event.exception.values || !hint || !isInstanceOf(hint.originalException, Error)) {
-      return event;
-    }
-
-    const linkedErrors = this._walkErrorTree(stackParser, hint.originalException as ExtendedError, this._key);
-    event.exception.values = [...linkedErrors, ...event.exception.values];
-    return event;
-  }
-
-  /**
-   * @inheritDoc
-   */
-  private _walkErrorTree(
-    stackParser: StackParser,
-    error: ExtendedError,
-    key: string,
-    stack: Exception[] = [],
-  ): Exception[] {
-    if (!isInstanceOf(error[key], Error) || stack.length + 1 >= this._limit) {
-      return stack;
-    }
-
-    const exception = exceptionFromError(stackParser, error[key]);
-
-    return this._walkErrorTree(stackParser, error[key], key, [exception, ...stack]);
   }
 }
